refactor(authz-rest): move AuthorizationConfig to plugin context (AIM pattern)
Move AuthorizationConfig out of the main rest-api Spring context and into
a dedicated plugin module context, matching how AIM's CatalogConfiguration
is wired. Eliminates the boot-time race where managementMongoTemplate was
not yet registered when AuthorizationConfig tried to resolve it.

Changes:

- Delete GammaSpringConfiguration (gamma-rest-api) and GammaAuthzSpringConfiguration
  (gamma-authorization-rest). Their sole purpose was the @Import chain into
  AuthorizationConfig.
- Drop GammaSpringConfiguration.class from StandaloneConfiguration's @Import.
  Main rest-api context no longer carries authz beans.
- Delete AuthorizationConfig from gamma-authorization-core. Its @Bean definitions
  move into the plugin module's SpringConfig (see below). gamma-authorization-core
  retains domain types, services, Mongo adapters, ports — just no Spring wiring.
- Add new gravitee-gamma-module-authz plugin module (id=authz, type=gamma-module).
  Its SpringConfig defines all the @Bean factories that used to live in
  AuthorizationConfig — repositories, services, listeners, event publisher, audit
  port. Living in the plugin jar means the JVM resolves @Bean parameter types
  (notably MongoOperations) through the plugin's URLClassLoader, whose parent is
  the apim-mongo plugin classloader where spring-data-mongodb actually sits.
- Bundle gamma-authorization-core inside the plugin zip via plugin-assembly.xml.
  Without it the plugin URLClassLoader loads MongoAuthz{Entity,Policy}Repository
  through graviteeClassLoader (lib/) and they fail with NoClassDefFoundError on
  spring-data-mongodb internals. Mark gamma-authorization-core as provided in
  gamma-authorization-rest so it stops getting pulled into rest-api dist's lib/
  on the way in — two copies of the same class on different classloaders cause
  ClassCastException across boundaries.
- @Qualifier on @Lazy @Bean params for the in-plugin repos/services: when
  GammaModulePluginHandler.initPluginSpringContext propagates plugin beans into
  the parent ctx with an "authz." prefix, the plugin ctx itself ends up seeing
  both copies via parent-ctx visibility — autowire-by-type would then fail with
  NoUniqueBeanDefinitionException at @Lazy proxy resolution time. Pinning by
  name picks the local one.
- Drop the @Lazy workaround on @Bean authzEntityRepository / authzPolicyRepository
  (and on their MongoOperations parameter) that landed on authz-backend. With
  AuthorizationConfig now in plugin context, the apim-mongo race the @Lazy was
  guarding against can no longer happen — the plugin context is created strictly
  after apim-mongo has registered managementMongoTemplate.

URL scheme unchanged: GammaRootResource keeps the hardcoded @Path("/authz")
subresource locator, AuthzRootResource and the exception mappers remain
registered by GammaModuleApplication. Resources still live in
gamma-authorization-rest (still a transitive dep of gamma-rest-api); only
the Spring wiring moved.
