+++ gravitee-apim-rest-api/gravitee-apim-rest-api-standalone/gravitee-apim-rest-api-standalone-container/src/main/java/io/gravitee/rest/api/standalone/spring/StandaloneConfiguration.java
@@ -15,7 +15,6 @@
  */
 package io.gravitee.rest.api.standalone.spring;
 
-import io.gravitee.gamma.rest.spring.GammaSpringConfiguration;
 import io.gravitee.integration.controller.spring.IntegrationControllerConfiguration;
 import io.gravitee.node.api.Node;
 import io.gravitee.node.api.NodeMetadataResolver;
@@ -44,7 +43,6 @@ import org.springframework.context.annotation.Import;
         io.gravitee.rest.api.management.v2.rest.spring.RestManagementConfiguration.class,
         IntegrationControllerConfiguration.class,
         RestPortalConfiguration.class,
-        GammaSpringConfiguration.class,
     }
 )
 public class StandaloneConfiguration {
