+++ gravitee-gamma/gravitee-gamma-module-authz/src/main/resources/plugin.properties
@@ -1,0 +1,6 @@
+id=authz
+name=Authorization Module
+type=gamma-module
+version=${project.version}
+description=Authorization module for Gravitee Gamma
+class=io.gravitee.gamma.module.authz.AuthorizationModule
