+++ gravitee-gamma/pom.xml
@@ -76,8 +76,9 @@
     <modules>
         <module>gravitee-gamma-plugin</module>
         <module>gravitee-gamma-rest-api</module>
+        <module>gravitee-gamma-authorization</module>
         <module>gravitee-gamma-module-apim</module>
         <module>gravitee-gamma-module-platform</module>
-        <module>gravitee-gamma-authorization</module>
+        <module>gravitee-gamma-module-authz</module>
     </modules>
 </project>
